+++ WBooks/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.awesomeproject",
+    package = "wbooks",
 )
 
 android_resource(
     name = "res",
-    package = "com.awesomeproject",
+    package = "wbooks",
     res = "src/main/res",
 )
