+++ WBooks/android/app/src/main/java/com/wbooks/MainActivity.java
@@ -1,4 +1,4 @@
-package com.awesomeproject;
+package com.wbooks;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "AwesomeProject";
+    return "WBooks";
   }
 }
