+++ WBooks/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AwesomeProject",
-  "displayName": "AwesomeProject"
+  "name": "WBooks",
+  "displayName": "WBooks"
 }
