+++ WBooks/ios/WBooks-tvOS/Info.plist
