+++ WBooks/ios/WBooks.xcodeproj/xcshareddata/xcschemes/WBooks.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "AwesomeProject.app"
-               BlueprintName = "AwesomeProject"
-               ReferencedContainer = "container:AwesomeProject.xcodeproj">
+               BuildableName = "WBooks.app"
+               BlueprintName = "WBooks"
+               ReferencedContainer = "container:WBooks.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AwesomeProjectTests.xctest"
-               BlueprintName = "AwesomeProjectTests"
-               ReferencedContainer = "container:AwesomeProject.xcodeproj">
+               BuildableName = "WBooksTests.xctest"
+               BlueprintName = "WBooksTests"
+               ReferencedContainer = "container:WBooks.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AwesomeProjectTests.xctest"
-               BlueprintName = "AwesomeProjectTests"
-               ReferencedContainer = "container:AwesomeProject.xcodeproj">
+               BuildableName = "WBooksTests.xctest"
+               BlueprintName = "WBooksTests"
+               ReferencedContainer = "container:WBooks.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeProject.app"
-            BlueprintName = "AwesomeProject"
-            ReferencedContainer = "container:AwesomeProject.xcodeproj">
+            BuildableName = "WBooks.app"
+            BlueprintName = "WBooks"
+            ReferencedContainer = "container:WBooks.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeProject.app"
-            BlueprintName = "AwesomeProject"
-            ReferencedContainer = "container:AwesomeProject.xcodeproj">
+            BuildableName = "WBooks.app"
+            BlueprintName = "WBooks"
+            ReferencedContainer = "container:WBooks.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeProject.app"
-            BlueprintName = "AwesomeProject"
-            ReferencedContainer = "container:AwesomeProject.xcodeproj">
+            BuildableName = "WBooks.app"
+            BlueprintName = "WBooks"
+            ReferencedContainer = "container:WBooks.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
