+++ WBooks/ios/WBooks/Images.xcassets/AppIcon.appiconset/Contents.json
