+++ WBooks/ios/WBooks/Images.xcassets/Contents.json
