+++ WBooks/ios/WBooksTests/Info.plist
