+++ WBooks/ios/WBooksTests/WBooksTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface AwesomeProjectTests : XCTestCase
+@interface WBooksTests : XCTestCase
 
 @end
 
-@implementation AwesomeProjectTests
+@implementation WBooksTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
